+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,56 @@
 # @sketch-hq/sketch-assistant-cli
 
+## 6.0.0
+
+### Major Changes
+
+- f8a0829: Support for reporting violations that reference multiple Sketch file objects.
+- 81a6e61: Add timeouts to rules.
+- 0b479bc: Refactored violation `locations` to `objects` and added a `class` value.
+
+### Patch Changes
+
+- 0aff579: Fix build error.
+- Updated dependencies [f8a0829]
+- Updated dependencies [81a6e61]
+- Updated dependencies [f155f44]
+- Updated dependencies [0b479bc]
+  - @sketch-hq/sketch-assistant-types@6.0.0
+  - @sketch-hq/sketch-assistant-utils@6.0.0
+
+## 6.0.0-rc.2
+
+### Patch Changes
+
+- 0aff579: Fix build error.
+
+## 6.0.0-rc.1
+
+### Major Changes
+
+- 0b479bc: Refactored violation `locations` to `objects` and added a `class` value.
+
+### Patch Changes
+
+- Updated dependencies [0b479bc]
+  - @sketch-hq/sketch-assistant-types@6.0.0-rc.1
+  - @sketch-hq/sketch-assistant-utils@6.0.0-rc.1
+
+## 6.0.0-rc.0
+
+### Major Changes
+
+- f8a0829: Support for reporting violations that reference multiple Sketch file objects.
+- 81a6e61: Add timeouts to rules.
+
+### Patch Changes
+
+- Updated dependencies [f8a0829]
+- Updated dependencies [81a6e61]
+- Updated dependencies [f155f44]
+  - @sketch-hq/sketch-assistant-types@6.0.0-rc.0
+  - @sketch-hq/sketch-assistant-utils@6.0.0-rc.0
+
 ## 5.0.0
 
 ### Major Changes